// src/balloons/types.ts
export type Balloon = {
    x: number;
    y: number;
    radiusX: number;
    radiusY: number;
    color: [number, number, number];
    driftPhase: number;
    driftFreq: number;
    driftAmp: number;
    riseSpeed: number;
    rotation: number;
    rotationSpeed: number;
    stringLength: number;
};